+++ frontend/src/app/accounts/login/page.tsx
@@ -1,4 +1,4 @@
-import LoginForm from '../../components/LoginForm'
+import LoginForm from '../../components/login-signup/LoginForm'
 export default function Page() {
     return (
         <div>
