+++ frontend/src/app/accounts/signup/page.tsx
@@ -1,4 +1,4 @@
-import SignupForm from '../../components/SignupForm'
+import SignupForm from '../../components/login-signup/SignupForm'
 export default function Page() {
     return (
         <div>
